+++ Methods/Queues/DELETE/removeFromArtistQueue.js
@@ -22,17 +22,15 @@ router.delete('/', async (req, res) => {
 	});
 
 	queues.findOne({ artistID: artistid }).then((result) => {
-		if (result.userids.length === 0) {
-			if (result.userids.length === 0) {
-				const fetchOptions = {
-					method: 'DELETE',
-					headers: {
-						queueid: result._id,
-						state: 'Artist',
-					},
-				};
-				fetch(`http://localhost:${process.env.PORT}/removeQueuefromDB`, fetchOptions);
-			}
+		if (result.userids && result.userids.length === 0) {
+			const fetchOptions = {
+				method: 'DELETE',
+				headers: {
+					queueid: result._id,
+					state: 'Artist',
+				},
+			};
+			fetch(`http://localhost:${process.env.PORT}/removeQueuefromDB`, fetchOptions);
 		}
 	});
 });
