+++ Methods/Queues/DELETE/removeFromSongQueue.js
@@ -23,7 +23,7 @@ router.delete('/', async (req, res) => {
 	});
 
 	queues.findOne({ songID: songid }).then((result) => {
-		if (result.userids.length === 0) {
+		if (result.userids && result.userids.length === 0) {
 			const fetchOptions = {
 				method: 'DELETE',
 				headers: {
